Update README.md, modularize and lint code
Pre-commit checks:
All checks passed ✅
