+++ research/agentic_data_science/schema_agent/README.md
@@ -12,25 +12,40 @@ Automated statistical profiling and LLM-powered semantic analysis for CSV datase
 
 ## Setup
 
-Go into the schema folder: 
-```bash 
-> cd research/agentic_data_science/schema_agent
+Go into the schema folder:
+```bash
+cd research/agentic_data_science/schema_agent
 ```
 
-Install the requirements with the command: 
+Install the requirements:
 ```bash
-> pip install -r requirements.txt
+pip install -r requirements.txt
 ```
-Set the OPENAI_API_KEY in the .env file: 
-```bash 
-> export OPENAI_API_KEY=sk-...
+
+Set the `OPENAI_API_KEY` in your environment:
+```bash
+export OPENAI_API_KEY=sk-...
 ```
+
+## Module Structure
+
+The agent is split into six focused modules:
+
+| Module | Responsibility |
+|--------|---------------|
+| `schema_agent_models.py` | Pydantic schemas for type-safe column/dataset insights |
+| `schema_agent_loader.py` | CSV loading, type inference, datetime detection |
+| `schema_agent_stats.py` | Numeric summaries, quality reports, categorical distributions |
+| `schema_agent_llm.py` | Prompt building, OpenAI/LangChain calls, structured output parsing |
+| `schema_agent_report.py` | Column profiles, JSON and Markdown export |
+| `schema_agent.py` | Pipeline orchestration and CLI entry point |
+
 ## Usage
 
 ### Basic
 
 ```bash
-python schema_agent_utils.py data.csv
+python schema_agent.py data.csv
 ```
 
 Outputs:
@@ -41,13 +56,16 @@ Outputs:
 
 ```bash
 # Multiple files with tags
-python schema_agent_utils.py dataset1.csv dataset2.csv --tags sales_2024 inv_q1
+python schema_agent.py dataset1.csv dataset2.csv --tags sales_2024 inv_q1
 
 # Cost-optimized: only high-null columns
-python schema_agent_utils.py data.csv --llm-scope nulls --model gpt-4o-mini
+python schema_agent.py data.csv --llm-scope nulls --model gpt-4o-mini
 
 # Custom metrics and output
-python schema_agent_utils.py data.csv --metrics mean std max --output-json my_report.json
+python schema_agent.py data.csv --metrics mean std max --output-json my_report.json
+
+# LangChain backend
+python schema_agent.py data.csv --use-langchain
 ```
 
 ## Command-Line Arguments
@@ -71,14 +89,26 @@ python schema_agent_utils.py data.csv --metrics mean std max --output-json my_re
 
 ## Python API
 
-```python
-import research.agentic_data_science.schema_agent.schema_agent_utils as radssasau
+### Full pipeline
 
-tag_to_df, stats = radssasau.run_pipeline(
+```python
+import schema_agent as radsasag
+tag_to_df, stats = radsasag.run_pipeline(
     csv_paths=["data.csv"],
     model="gpt-4o-mini",
     llm_scope="semantic"
 )
+```
+
+### Individual modules
+
+Each module can be imported independently for exploratory use or testing:
+
+```python
+import schema_agent_loader as radsasal
+import schema_agent_stats as radsasas
+import schema_agent_llm as radsasal
+import schema_agent_report as radsasar
 ```
 
 ## Output
